Checkout the website mentioned below
RAPIDOC
RAPIDOC is an online platform based on providing healthcare facilites to the user and information regarding hospital like availability of beds, total number of patients,,helping user to prebook appointments for OPD ,check availability of blood groups in hospitals and many more presentable in next slides. Its aim is to improve access to quality healthcare and addressing health challenges.It helps to promote betterment of society.
So basically the benefits which RAPIDOC will provide you are as follows:
1.Locating Hospital and Clinics
It is difficult for us to differentiate by seeing on map whether the hospital is government,private or a clinic so to overcome this problem we basically provide you the interface of map where you can checkout your closest Hospital and clinic  with differentiability available.
2.Online Registeration
Now talking about online registration it is particularly based on (Outpatient department)OPD section in hospitals.Usually in government or private hospitals their is so long waiting for the checkup a suffering patient has too line up for it and after waiting for hours he/she able to get his/her checkup done.So to overcome this we will provide online registration for the OPD section.
3.Availiability
It is difficult for us to see whether the beds are available in the hospital or not irrespective of being present their physical.So in this section we will provide the number of beds available in the hospital so that if any emergency case would be there one could not suffer from lack of availability at that particular time.Also we will provide the blood sample available in the hospitals to make it easy to check the availability if any hospital is lagging with it.
4.Pharmacy
RAPIDOC will provide you the 24*7 service for providing medicine,medical equipment,machines like Blood glucose monitor and many more.
5.Balance Life 
Nowadays every next person is suffering from some chronic disease like diabetes,thyroid,cholesterol,etc and many more.
So in this we will provide you how are the symptoms of a particular disease,precautionary measure you should take if you are suffering from it, Yoga which helps you to overcome it,Food preferred to eat at that time period.
6.BMI Calculator
It helps to get in check with your weight according to your height.It helps you to track your physical activity in day to day life like steps walk and how much calories burn also helps to know heart rate.