Update README.md
This pull request adds detailed instructions for contributors on how to contribute to the RAPIDOC-HEALTHCARE-WEBSITE repository using GitHub Desktop. The instructions cover the process of forking the repository, cloning it to the local machine, making changes, committing them, pushing them to the forked repository, and creating a pull request. The guide aims to make it easier for new contributors to understand and follow the contribution process using GitHub Desktop.