Add files via upload
I have corrected the header and footer of the website. I have also created a favicon icon, which is the logo of the website near the title of the page. Also, i have corrected the positioning of the looping video in the right side which was interefering with the alignment of the navbar. Also i have added spaces between all the logos and the names of button in the navbar. And some other aligning here and there have also been rectified.